Clarify series metadata API w/in README.md (#26)
* Clarify README.md

I was worried to run the last example because I was worried it would generate a lot of bandwidth. I clarified what the example does. Also, when copy-pasting, I forgot to change series name in one place and I was getting a strange error. I decided to factor it out to a separate variable so nobody else would fall for this.

* s/metadata all/metadata for all/g